+++ client/src/components/Salary/StyledElements.ts
@@ -11,7 +11,7 @@ export const StyledTableCell = styled(TableCell)(
       color: theme.palette.common.white,
     },
     [`&.${tableCellClasses.body}`]: {
-      fontSize: 14,
+      fontSize: 16,
     },
   })
 );
